+++ CHANGELOG.md
@@ -4,6 +4,17 @@
 
 No date
 
+## [0.21.0]
+
+20.07.2021
+
+-   Add French translation (thanks @Darkomen78).
+-   Fix version for which `--allowremoval` is set when doing a reinstall (thanks @anverhousseini).
+-   Kill DEPNotify in places where jamfHelper is killed (#106, thanks @julienvs).
+-   Added '$script_name' variable - if you want to change the script name, the echo statements will reflect this value rather than 'erase-install'.
+-   Added `--clear-cache-only` option, which works in conjunction with `--overwrite` or `--update` to perform the removal of cached installers but then quit rather than carry on with any further workflow (#105).
+-   Added a more verbose message that `--test-run` has been implemented (#93).
+
 ## [0.20.1]
 
 12.05.2021
@@ -310,7 +321,8 @@ Thanks to '@ahousseini' for various contributions to this release
 
 -   Initial version. Expects a manual choice of installer from `installinstallmacos.py`.
 
-[untagged]: https://github.com/grahampugh/erase-install/compare/v0.20.1...HEAD
+[untagged]: https://github.com/grahampugh/erase-install/compare/v0.21.0...HEAD
+[0.21.0]: https://github.com/grahampugh/erase-install/compare/v0.20.1...v0.21.0
 [0.20.1]: https://github.com/grahampugh/erase-install/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/grahampugh/erase-install/compare/v0.19.2...v0.20.0
 [0.19.2]: https://github.com/grahampugh/erase-install/compare/v0.19.1...v0.19.2